backend/requirements.txt
Flask==3.1.0
Flask-CORS==5.0.1
Flask-SQLAlchemy==3.1.1
PyJWT==2.10.1
lmstudio>=1.0.0
python-dotenv==1.1.0
